Extract VespaAssertions module from Assertions
Move assert_json_string_equal and assert_tensors_equal into a new
VespaAssertions module, separating Vespa-specific assertions from the
generic Test::Unit-style Assertions module.
